user/lib: Introduce custom test runner
Introduce a custom test runner in `userlib` for executing tests in
userspace, following the same pattern used in the kernel.

Enable it via the `test_runner` feature, which also brings the custom
`test` crate from the workspace only when testing with the
`x86_64-unknown-none` target. If the feature is enabled for a different
target, Cargo will ignore the custom `test` crate.

This is because there is a conflict with the `test` crate during normal
testing, resulting in the  custom crate being imported instead of the
standard one. This causes the following error when running Clippy:

```
$ cargo clippy --all-features --package userlib --tests
    Checking test v0.1.0 (...)
    ...
    Checking userlib v0.1.0 (...)
error[E0425]: cannot find function `test_main_static` in crate `test`
```

Signed-off-by: Nicola Ramacciotti <niko.ramak@gmail.com>