user/lib: Introduce custom test runner
Introduce a custom test runner in `userlib` for executing tests
in userspace, following the same pattern used in the kernel.
Enable it via the `test_runner` feature, which also brings the
custom `test` crate from the workspace.

Exclude the new `test_runner` feature from clippy checks since
it is only used for `test_in_svsm`, not for regular tests that
use `--all-features`.

Signed-off-by: Nicola Ramacciotti <niko.ramak@gmail.com>